Revert "Add teacher create test"
