+++ ATlearning/ATframework3demo/BaseFramework/TestCase.cs
@@ -43,7 +43,6 @@ namespace atFrameWork2.BaseFramework
                 if (EnvType == TestCaseEnvType.Web)
                 {
                     var scheduleLoginPage = new ScheduleLoginPage();
-                    WebDriverActions.OpenUrl("http://project/login/");
                     var homePage = scheduleLoginPage.Login(testPortal.PortalAdmin);
                     Body.Invoke(homePage);
                 }
