+++ ATlearning/ATframework3demo/TestCases/Case_SomeCase.cs
@@ -1,0 +1,21 @@
+﻿using atFrameWork2.BaseFramework;
+using atFrameWork2.BaseFramework.LogTools;
+using atFrameWork2.PageObjects;
+using atFrameWork2.SeleniumFramework;
+
+namespace ATframework3demo.TestCases
+{
+    public class Case_SomeCase : CaseCollectionBuilder
+    {
+        protected override List<TestCase> GetCases()
+        {
+            var caseCollection = new List<TestCase>();
+            caseCollection.Add(new TestCase("чето там", homePage => Test(homePage)));
+            return caseCollection;
+        }
+
+        public void Test(ScheduleHomePage homePage) {
+            return;
+        }
+    }
+}
